Feature addition and reformat of jap_words.csv
An additional command line feature was added to allow what is being
tested for each word to be specified. The hiragana of a word featuring
kanji will now be displayed when passing on a word.

The rework of the csv file merely involves removing the word type from
the list of word parameters, felt it was unecessary and would merely
slow the addition of other words.